moved test to LastestLoaderIT
